把文件中大部分int改为long long
Maybe the user wants to deal with long long type number,so I changed this.
也许用户会用此处理long long 数据，所以我改了这个

+++ docs/math/code/inverse/inverse-2.cpp
@@ -2,7 +2,7 @@
 
 // --8<-- [start:core]
 // Binary exponentiation.
-int pow(int a, int b, int m) {
+long long pow(long long a, long long b, long long m) {
   long long res = 1, po = a;
   for (; b; b >>= 1) {
     if (b & 1) res = res * po % m;
@@ -12,14 +12,14 @@ int pow(int a, int b, int m) {
 }
 
 // Returns the modular inverse of a prime modulo p.
-int inverse(int a, int p) { return pow(a, p - 2, p); }
+long long inverse(long long a, long long p) { return pow(a, p - 2, p); }
 
 // --8<-- [end:core]
 int main() {
-  int t;
+  long long t;
   std::cin >> t;
   for (; t; --t) {
-    int a, p;
+    long long a, p;
     std::cin >> a >> p;
     std::cout << inverse(a, p) << std::endl;
   }
